Exception handling - delete

+++ src/main/java/com/joao/project/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.joao.project.entities.User;
 import com.joao.project.repositories.UserRepository;
+import com.joao.project.services.exceptions.DatabaseException;
 import com.joao.project.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -31,7 +33,16 @@ public class UserService {
     }
 
     public void deleteById(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException(id);
+        }
+        try{
+            repository.deleteById(id);
+        }
+        catch(DataIntegrityViolationException e){
+            throw new DatabaseException(e.getMessage());
+        }
+     
     }
     public User update(Long id, User obj){
         User entity = repository.getReferenceById(id);
